+++ Arcanoid/Assets/Resources/Prefabs/Ball.prefab
@@ -12,6 +12,7 @@ GameObject:
   - component: {fileID: 8975714604946854907}
   - component: {fileID: 2088714276932577726}
   - component: {fileID: 656296288783413426}
+  - component: {fileID: 3339366643503450481}
   m_Layer: 0
   m_Name: Ball
   m_TagString: Untagged
@@ -117,3 +118,19 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 04a07a432eaaaa34d84224040da0c15b, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+--- !u!58 &3339366643503450481
+CircleCollider2D:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 4111507523595127104}
+  m_Enabled: 1
+  m_Density: 1
+  m_Material: {fileID: 0}
+  m_IsTrigger: 0
+  m_UsedByEffector: 0
+  m_UsedByComposite: 0
+  m_Offset: {x: 0, y: 0}
+  serializedVersion: 2
+  m_Radius: 0.5
