+++ Arcanoid/Assets/Scripts/BoardBuilder.cs
@@ -30,9 +30,24 @@ namespace Arcanoid
             {
                 _tileFactory.GetTileAt(point, _cellSize);
             }
+            SpawnPlayer();
+
         }
 
+        private void SpawnPlayer()
+        {
+            var boarPosition = GetBoardSpawnPoint();
+            var boardScale = new Vector3(_gameSettings.BoardWidth, 0.3f, 1f);
+            _player.GetTileAt(boarPosition, boardScale);
 
+        }
+
+        private Vector3 GetBoardSpawnPoint()
+        {
+            var x = (Mathf.Abs(_screenBounds.TopRight.x) - Mathf.Abs(_screenBounds.BottomLeft.x))/2;
+            var y = 1f; // magick number
+            return new Vector3(x, y, 0f);
+        }
 
         private void GetSpawnPoints()
         {
