+++ Arcanoid/Assets/Scripts/BoardFactory.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4c65e6df37493d34595a7f0f7ab22b8f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
