+++ Arcanoid/Assets/Scripts/GameStarter.cs
@@ -14,10 +14,6 @@ namespace Arcanoid
             var settings = Resources.Load<GameSettings>($"Settings/GameSettings");
             var screenBounds = new ScreenBounds(Camera.main);
             
-
-            var board = FindObjectOfType<BoardTile>();
-            board.SetScale(new Vector3(settings.BoardWidth, 0.3f,1f));
-            var ball = FindObjectOfType<Ball>();
             var input = FindObjectOfType<PCInput>();
 
             var boardSpeedModel = new SpeedModel(settings.BoardSpeed);
@@ -25,22 +21,18 @@ namespace Arcanoid
             var hitModel = new HitModel(settings.States);
             var hitViewModel = new HitViewModel(hitModel);
 
-            var tileFactory = new TileFactory(hitModel, hitViewModel, settings);
-
-            var boardFiller = new BoardBuilder(screenBounds, settings, tileFactory);
-            boardFiller.FillBoard();
-
-
-
             var boardSpeedViewModel = new MovementViewModel(boardSpeedModel);
             var ballSpeedViewModel = new MovementViewModel(ballSpeedModel);
             var inputVM = new InputViewModel();
 
+            var tileFactory = new TileFactory(hitModel, hitViewModel, settings);
+            var playerFactory = new BoardFactory(settings, inputVM, boardSpeedViewModel, ballSpeedViewModel);
+
+
+            var boardFiller = new BoardBuilder(screenBounds, settings, tileFactory, playerFactory);
+            boardFiller.FillBoard();
+
             input.Construct(inputVM);
-            ball.Construct(inputVM, ballSpeedViewModel);
-            board.Construct(inputVM, boardSpeedViewModel);
-
-
         }
 
     }
