+++ Arcanoid/Logs/shadercompiler-UnityShaderCompiler.exe0.log
@@ -1,3 +1,3 @@
-Base path: 'C:/Program Files/Unity/Hub/Editor/2020.3.12f1/Editor/Data', plugins path 'C:/Program Files/Unity/Hub/Editor/2020.3.12f1/Editor/Data/PlaybackEngines'
+Base path: 'G:/Unity/2020.3.12f1/Editor/Data', plugins path 'G:/Unity/2020.3.12f1/Editor/Data/PlaybackEngines'
 Cmd: initializeCompiler
